pyproject.toml
[tool.poetry]
name = "sarah-streamlit"
version = "0.1.0"
description = "A modern Streamlit chat application with Claude integration via Vertex AI"
authors = ["Joshua Yeomans"]
readme = "README.md"
packages = [{include = "sarah_streamlit", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.32.0"
pandas = "^2.2.0"
numpy = "^1.26.3"
google-cloud-aiplatform = "^1.45.0"
anthropic = "^0.45.2"
python-dotenv = "^1.0.1"
supabase = "^2.13.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
ruff = "^0.2.1"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C", "D"]
ignore = ["D100", "D104"]

[tool.ruff.pydocstring]
convention = "google" 